Commit
Commit

+++ StepTest/app/src/main/java/com/paobuqianjin/pbq/step/data/bean/gson/response/DynamicIdDetailResponse.java
@@ -72,7 +72,7 @@ public class DynamicIdDetailResponse  {
         private String city;
         private int vote;
         private int comment;
-        private int create_time;
+        private long create_time;
         private List<String> images;
 
         public int getId() {
@@ -139,11 +139,11 @@ public class DynamicIdDetailResponse  {
             this.comment = comment;
         }
 
-        public int getCreate_time() {
+        public long getCreate_time() {
             return create_time;
         }
 
-        public void setCreate_time(int create_time) {
+        public void setCreate_time(long create_time) {
             this.create_time = create_time;
         }
